Refactor Lumen Threat Feed Functions
- Removed the main function app code from function_app.py, separating concerns into individual activity functions for cleanup and upload.
- Introduced activity functions: activity_cleanup_blob and activity_upload_from_blob for better modularity and maintainability.
- Enhanced orchestrator_function to dynamically distribute work units and track progress more effectively.
- Updated starter_function and timer_starter_function to utilize the new orchestration model.
- Adjusted host.json to simplify configuration and removed unnecessary durable task settings.
- Updated deployment template to reflect changes in function URLs and added necessary app settings for blob storage.
- Improved logging throughout the functions for better traceability and debugging.
