+++ Solutions/Lumen/Data Connectors/LumenThreatFeed/LumenThreatFeedV4/.funcignore
@@ -1,2 +1,3 @@
 .venv
 .vscode
+local.settings.json
