+++ Solutions/Lumen/Data Connectors/LumenThreatFeed/LumenThreatFeedV4/.gitignore
@@ -57,4 +57,4 @@ logs/
 .python_packages
 
 #vscode
-.vscode/
+.vscode
