+++ Solutions/Lumen/Data Connectors/LumenThreatFeed/LumenThreatFeedV4/activity_cleanup_blob/function.json
@@ -1,0 +1,10 @@
+{
+  "scriptFile": "__init__.py",
+  "bindings": [
+    {
+      "name": "cleanup_info",
+      "type": "activityTrigger",
+      "direction": "in"
+    }
+  ]
+}
