+++ Solutions/Lumen/Data Connectors/LumenThreatFeed/LumenThreatFeedV4/activity_upload_from_blob/function.json
@@ -1,0 +1,10 @@
+{
+  "scriptFile": "__init__.py",
+  "bindings": [
+    {
+      "name": "work_unit",
+      "type": "activityTrigger",
+      "direction": "in"
+    }
+  ]
+}
