+++ Solutions/Lumen/Data Connectors/LumenThreatFeed/LumenThreatFeedV4/host.json
@@ -12,15 +12,5 @@
     "id": "Microsoft.Azure.Functions.ExtensionBundle",
     "version": "[4.*, 5.0.0)"
   },
-  "functionTimeout": "00:20:00",
-  "extensions": {
-    "durableTask": {
-      "hubName": "LumenThreatFeedV4Hub",
-      "maxConcurrentActivityFunctions": 10,
-      "maxConcurrentOrchestratorFunctions": 1,
-      "storageProvider": {
-        "connectionStringName": "AzureWebJobsStorage"
-      }
-    }
-  }
+  "functionTimeout": "00:10:00"
 }
