+++ Solutions/Lumen/Data Connectors/LumenThreatFeed/LumenThreatFeedV4/orchestrator_function/function.json
@@ -1,0 +1,10 @@
+{
+  "scriptFile": "__init__.py",
+  "bindings": [
+    {
+      "name": "context",
+      "type": "orchestrationTrigger",
+      "direction": "in"
+    }
+  ]
+}
