+++ Solutions/Lumen/Data Connectors/LumenThreatFeed/azuredeploy_Connector_LumenThreatFeed_V4.json
@@ -164,8 +164,8 @@
           "appSettings": [
             { "name": "FUNCTIONS_EXTENSION_VERSION", "value": "~4" },
             { "name": "FUNCTIONS_WORKER_RUNTIME", "value": "python" },
-            { "name": "PYTHON_ISOLATE_WORKER_DEPENDENCIES", "value": "1" },
             { "name": "AzureWebJobsStorage", "value": "[concat('DefaultEndpointsProtocol=https;AccountName=', variables('storageAccountName'), ';AccountKey=', listKeys(resourceId('Microsoft.Storage/storageAccounts', variables('storageAccountName')), variables('storageApiVersion')).keys[0].value, ';EndpointSuffix=', environment().suffixes.storage)]" },
+            { "name": "WEBSITE_RUN_FROM_PACKAGE", "value": "https://raw.githubusercontent.com/lumen-mcollier/Azure-Sentinel/master/Solutions/Lumen/Data%20Connectors/LumenThreatFeed/LumenThreatFeedV4.zip" },
             { "name": "LUMEN_API_KEY", "value": "[parameters('LUMEN_API_KEY')]" },
             { "name": "LUMEN_BASE_URL", "value": "[parameters('LUMEN_BASE_URL')]" },
             { "name": "WORKSPACE_ID", "value": "[parameters('WORKSPACE_ID')]" },
@@ -178,7 +178,8 @@
             { "name": "LUMEN_ENABLE_DOMAIN", "value": "true" },
             { "name": "APPINSIGHTS_INSTRUMENTATIONKEY", "value": "[reference(resourceId('Microsoft.Insights/components', variables('appInsightsName')), '2020-02-02').InstrumentationKey]" },
             { "name": "APPLICATIONINSIGHTS_CONNECTION_STRING", "value": "[reference(resourceId('Microsoft.Insights/components', variables('appInsightsName')), '2020-02-02').ConnectionString]" },
-            { "name": "WEBSITE_RUN_FROM_PACKAGE", "value": "https://raw.githubusercontent.com/lumen-mcollier/Azure-Sentinel/raw/refs/heads/master/Solutions/Lumen/Data%20Connectors/LumenThreatFeed/LumenThreatFeedV4.zip" }
+            { "name": "WEBSITE_CONTENTAZUREFILECONNECTIONSTRING", "value": "[concat('DefaultEndpointsProtocol=https;AccountName=', variables('storageAccountName'), ';AccountKey=', listKeys(resourceId('Microsoft.Storage/storageAccounts', variables('storageAccountName')), variables('storageApiVersion')).keys[0].value, ';EndpointSuffix=', environment().suffixes.storage)]" },
+            { "name": "WEBSITE_CONTENTSHARE", "value": "[toLower(variables('uniqueFuncName'))]" }
           ]
         }
       }
@@ -215,7 +216,7 @@
     },
     "ManualTriggerUrl": {
       "type": "string",
-      "value": "[concat('https://', variables('uniqueFuncName'), '.azurewebsites.net/api/starter')]",
+      "value": "[concat('https://', variables('uniqueFuncName'), '.azurewebsites.net/api/starter_function')]",
       "metadata": {
         "description": "URL to manually trigger the threat feed update"
       }
